Add map[string][]*Route
